Extracting building logic into a workers package.
