+++ cmd/gaia/main.go
@@ -12,8 +12,8 @@ import (
 	"github.com/gaia-pipeline/flag"
 	"github.com/gaia-pipeline/gaia"
 	"github.com/gaia-pipeline/gaia/handlers"
-	"github.com/gaia-pipeline/gaia/pipeline"
-	"github.com/gaia-pipeline/gaia/services"
+	"github.com/gaia-pipeline/gaia/workers/pipeline"
+	"github.com/gaia-pipeline/gaia/workers/services"
 	hclog "github.com/hashicorp/go-hclog"
 	"github.com/labstack/echo"
 )
