+++ handlers/pipeline_test.go
@@ -16,7 +16,7 @@ import (
 	"github.com/gaia-pipeline/gaia/services"
 
 	"github.com/gaia-pipeline/gaia"
-	"github.com/gaia-pipeline/gaia/pipeline"
+	"github.com/gaia-pipeline/gaia/workers/pipeline"
 	hclog "github.com/hashicorp/go-hclog"
 	"github.com/labstack/echo"
 )
