+++ workers/pipeline/update_pipeline_test.go
